+++ MergeSort/MergeSortTest.cs
@@ -24,7 +24,7 @@ namespace MergeSort
             var left = new[] {1, 3, 5};
             var right = new[] {2, 4, 6};
 
-            int splitInversionCount;
+            long splitInversionCount;
             var merged = MergeSort<int>.Merge(left, right, out splitInversionCount);
 
             CollectionAssert.IsOrdered(merged);
@@ -38,7 +38,7 @@ namespace MergeSort
         [TestCase(new[] { 1, 2, 3, 5, 4 }, Result = 1)]
         [TestCase(new[] { 3, 1, 6, 5, 2, 4 }, Result = 7)]
         [TestCase(new[] { 5, 2, 10, 8, 1, 9, 4, 3, 6, 7 }, Result = 22)]
-        public int SortsAndCountsInversionsCorrectly(int[] numbers)
+        public long SortsAndCountsInversionsCorrectly(int[] numbers)
         {
             IEnumerable<int> sorted;
             var inversions = MergeSort<int>.CountInversions(numbers, out sorted);
